refactor(functions): revert to old structure to remove circ dep
